Jest Tests added
