+++ .env
@@ -7,3 +7,5 @@ DB_USER = 'postgres'
 DB_DB = 'testOrders'
 DB_HOST = 'localhost'
 DB_PORT = '5432'
+
+TOKEN = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkF5dXNoIFZlcm1hIiwiaWF0IjoxNTE2MjM5MDIyfQ.56cj8MooW4zmg_O1h3CXlxe84D2aEPcjuXTF1i8sJLA'
