+++ package.json
@@ -15,10 +15,12 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "express-json-validator-middleware": "^3.0.1",
+    "jest": "^29.5.0",
     "jsonwebtoken": "^9.0.0",
     "moment": "^2.29.4",
     "nodemon": "^2.0.21",
     "pg": "^8.10.0",
-    "sequelize": "^6.29.3"
+    "sequelize": "^6.29.3",
+    "supertest": "^6.3.3"
   }
 }
